% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dNorway.R
\docType{data}
\name{dNorway}
\alias{dNorway}
\title{Example data set for Covid-19 data in Norway}
\format{
A list of 5 objects, each explained above
}
\usage{
data(dNorway)
}
\description{
A dataset containing the necessary data to run the SMC algorithm. The variables are as follows:
}
\details{
\itemize{
  \item hosp: A data.table with two columns, 'date' and 'N' where 'N' is the number of new hospital admissions at day 'date'
  \item test: A data.table with four columns, 'Negative', 'Positive'
  \item seed
  \item mob_mat
  \item pop
}
}
\keyword{datasets}
