% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_convert_test_data.R
\name{smc_convert_test_data}
\alias{smc_convert_test_data}
\title{Modify test data
Particle simulation of the spread model and log-likelhood evaluation based on hospital data 
Mainly to be used within smc_loglik
Global variables used:
'mobility_matrices', assumed to start at 'input_date'
'hospD'
'testD'}
\usage{
smc_convert_test_data(testD, date, pi.test = c(-1.175, 7.4e-05), week = FALSE)
}
\arguments{
\item{'testD'}{A data.frame/table with columns 'Negative', 'Positive' and 'date'}

\item{'pi'}{A vector giving pi_0 and pi_1 for specifying probability for positiv detection}

\item{'date'}{A vector for which dates to construct new test data object}
}
\value{
A data.frame with columns date and for every 7 day the sum of the previous positives and totals
}
\description{
First particles are 'unit.samp' days from 'input_date'
Thereafter log-likelhood is calculated based on hospital data
}
\details{
There are currently many global parameter used in the underlying routines. No control on that
}
