% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_extract_c.R
\name{smc_extract_c}
\alias{smc_extract_c}
\title{Extract c_symb_incidence+c_asymp_incidence summed by day from a (list of) data tables}
\usage{
smc_extract_c(particle)
}
\arguments{
\item{'particle'}{A data table containing at least c_symb_incidence and c_asymp_incidence}

\item{'particles'}{A list of particles, each a data table containing at least c_symb_incidence and c_asymp_incidence}
}
\value{
A vector/matrix with rows corresponding to days and if calling smc_extract_list, a column corresponding to particles
}
\description{
Extract c_symb_incidence+c_asymp_incidence summed by day from a (list of) data tables
}
