% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_init_sim.R
\name{smc_init_sim}
\alias{smc_init_sim}
\title{Initial simulation of the spread model 
Mainly to be used within smc_loglik}
\usage{
smc_init_sim(
  input,
  start_date,
  end_date,
  mob_mat,
  loc.seed,
  pop,
  MaxdaysH,
  pbar,
  MaxdaysT,
  qbar
)
}
\arguments{
\item{'input'}{A list with 'R0', 'AMP'}

\item{'start_date'}{start date for simulation}

\item{'end_date'}{end date for simulation}

\item{'mob_mat'}{Mobility matrix, supposed to start at start_date}

\item{'loc.seed'}{seeding for call to run_model}

\item{'pop'}{population structure}

\item{'MaxdaysH'}{maximum delay days for hospitalization}

\item{'pbar'}{Survival probability for day of hospitalization}

\item{'MaxdaysT'}{maximum delay days for testing}

\item{'qbar'}{Survival probability for day of testing}
}
\value{
A list containing 'particle', simulated particles of latent variables, 'insH', simulated numbers of potential Hospitalizations for each day (first index is infection day, second index is delay, starting at zero), 'insT' is delay of potential test with same structure as 'insH'
}
\description{
First initial particles are generated until an initial date
}
\details{
There are currently many global parameter used in the underlying run_model routine. No control on that
}
