% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_loglik_mem.R
\name{smc_loglik_mem}
\alias{smc_loglik_mem}
\title{Marginal loglikelihood estimation by sequential Monte Carlo}
\usage{
smc_loglik_mem(
  mu_R0 = 3.3,
  sd_R0 = 0.1,
  mu_AMP = 2.8,
  sd_AMP = 0,
  dyn_model = "RW",
  param_dynamicR = list(alpha = 2.4, beta = 0.28, phi.a = 100, phi.b = 900, a0 = 0.8,
    mu0 = 0, kappa0 = 1),
  lambda = 9.08,
  lambda_size = NULL,
  lambda_ch = NULL,
  prob_h = c(0.036, 0.022),
  change_prob = c(as.Date("2020-05-01")),
  alpha_hosp = NULL,
  alpha_test = NULL,
  q_test = c(0, 0.05, 0.08, 0.16, 0.4, 0.3, 0.01),
  pi_test = c(-1.175, 7.4e-05),
  start_date = as.Date("2020-02-17"),
  dynR_date = as.Date("2020-03-08"),
  end_date = as.Date("2020-05-05"),
  hosp_data = hospitalisations_df,
  test_data = NULL,
  pop = pop,
  loc.seed = seeding,
  mob_mat = mobility_matrices,
  resdir = "./",
  output.call = "call.txt",
  output.R = "smc_dynamic_R.txt",
  output.R_smo = "smc_dynamic_R_smo.txt",
  output.R_lag = "smc_dynamic_R_lag.txt",
  output.W = "smc_weights.txt",
  output.AMP = "smc_AMP.txt",
  output.SdR = "smc_SdR.txt",
  output.C = "smc_dynamic_C.txt",
  output.C_smo = "smc_dynamic_C_smo.txt",
  output.C_lag = "smc_dynamic_C_lag.txt",
  output.phat = "smc_phat.txt",
  B = 2000,
  Bmem = B,
  B2 = NULL,
  unit.samp = 3,
  resamp.frac = 1.1,
  smo.lag = 25,
  parallel = TRUE,
  pmcmc = FALSE,
  fix.R = NULL,
  sim.data = FALSE,
  VERBOSE = TRUE,
  DEBUG = FALSE,
  output_latent = FALSE,
  seed = 123,
  N_cores = 4
)
}
\arguments{
\item{mu_R0}{R0 in the first static period is assumed to follow a lognormal distribution with expectation mu_R0. Default is 3.3.}

\item{sd_R0}{The standard deviation in the lognormal distribution for R0. Default is 0.1.}

\item{mu_AMP}{R0 The amplification factor AMP is assumed to follow a lognormal distribution with expectation mu_AMP. Default is 2.8.}

\item{sd_AMP}{The standard deviation in the lognormal distribution for AMP. Default is 0.0 in which case no updating of AMP is performed.}

\item{dyn_model}{The model for dynamic R, currently only RW (random walk) is implemented}

\item{param_dynamicR}{List with parameters related to dynamic model for R. alpha,beta are shape and rate parameters for gamma distribution of precision, a,b are parameters in beta distribution for prob of changepoint, mu0,kappa0 are mean and proportional precision for mean}

\item{lambda}{Expected length from sickness to hospitalization. Default to 9.08.}

\item{lambda_size}{Size parameter in the Negative binomial distribution for length until hospitalization. Default is NULL in which case a Poisson distribution is used.}

\item{lambda_ch}{If 'lambda' and 'lambda_size' are vectors, this correspond to the dates for changes in these parameters}

\item{prob_h}{Probability for hospitalisation. Default to 0.036.}

\item{alpha_hosp}{First parameter in beta-binomial distribution for hospitalization. If NULL, the model reduces to a binomial distribution with 'prob_h' being the probability for hospitalization. In general, the second (beta) parameter is calculated as 'alpha_hosp * (1 - prob_h) / prob_h'. Default is NULL.}

\item{alpha_test}{First parameter in beta-binomial distribution for testdata. If NULL, the model reduces to a binomial distribution with 'prob_h' being the probability for hospitalization. In general, the second (beta) parameter is calculated as 'alpha_hosp * (1 - prob_h) / prob_h'. Default is NULL.}

\item{pi_test}{Detection probability for test data}

\item{start_date}{starting date for initial simulations}

\item{end_date}{final date for doing simulation (note that hospital data used after this date)}

\item{hosp_data}{data frame containing hospital data, with components date and N.}

\item{test_data}{data frame containing test data, with components ...If NULL, only hospital data is used. Can contain NA}

\item{pop}{data table describing population structure}

\item{mob_mat}{list of mobility matrices for each (6 hours) time-interval}

\item{output.R}{Filename for writing out dynamic R simulations}

\item{output.R_smo}{Filename for writing out dynamic R simulations smoothing backwards.}

\item{output.R_lag}{Filename for writing out dynamic R simulations smoothing 'smo.lag' units backwards.}

\item{output.C}{Filename for writing out c_symp+c_asymp}

\item{output.C_smo}{Smoothed version of previous}

\item{output.C_lag}{Lagged version of previous}

\item{output.phat}{Writing estimate of marginal likelihood to file}

\item{B}{Number of particles. Default to 20.}

\item{B2}{The storage of all the latent variable could be problematic due to memory. A smaller number of particles to be stored can be specified through B2 in which case only B2 particles are sampled at the end and returned. Default is NULL is which case no downscaling is performed.}

\item{unit.samp}{The number of days for which simulation is performed within each iteration (in the dynamic period). When call to the run_model routine is made within each iteration, some of the latent variables (mobility information) is not transfered, making the simulations somewhat approiximate. A large unit.samp value is preferred from this perspective. However, importance weights are calculated based on the number of days that are simulated, giving more extreme weights for a larger number of unit.samp. A trade-off is therefore needed. Values of 3 and 5 seems to work ok. Default is 3.}

\item{resamp.frac}{Resampling will increase variability at the current time (but might decrease later). When the effective sample size is smaller than B*resamp.frac, resampling is performed. For the current implementation a value larger than 1 is recommended, in which case resampling will always be performed. Default is 1.1.}

\item{smo.lag}{Fixed lag smoothing is based on inference from \eqn{p(x_t|y_{t+h})} where $h$ corresponds to 'smo.lag'. A large value increase the data information but due to resampling may leed to very few unique samples. Recommended (and default) value is 25.}

\item{parallel}{TRUE of parallelization during computation is to be used. Should always be used (if not in a debugging setting). Default is TRUE.}

\item{fix.R}{If non-null, contains an ndays x B matrix with fixed R's (so they are not simulated)}

\item{sim.data}{If TRUE, data are simulated, and replaces the input data}

\item{VERBOSE}{If TRUE, some output will be written during excecution. Default is TRUE}

\item{DEBUG}{If TRUE, extra information is written during excecution. Should only be used during debugging. Default is FALSE.}

\item{output_latent}{If FALSE, output of particles for latent variables is suppressed in order to save memory. Default is FALSE}

\item{seed}{Seed number. Default to 123. Note that the 'run_model' function that is called frequently do not use this seed, so that it is currently not possible to obtain a full repetition of the simulations.}

\item{init_date}{date until particles are simulated before calibration towards data}
}
\value{
A list containing particles of latent variables ('latent'), particles of parameters ('param') and estimate of marginal log-likelihood.
}
\description{
R is assumed to be constant until 'dynR_date', thereafter following a random walk distribution on log-scale
The variance for the random walk is estimated by using the method of sufficient statistics (Storvik-2002).
Sequential Monte Carlo is performed through the Bootstrap filter, that is the latent variables (particles) are simulated from their prior.
Importance weights are calculated based on hospital data which are assumed to follow a Beta-binomial distribution for hospitalization and
a Negative binomial distribution for the length from symptons to hospitalization.
}
\details{
Latent variables are simulated by the prior model (the spread model) using the run_model routine
First initial particles are generated until an initial date using the smc_init_sim routine.
Within the dynamic period, the particles are simulated using the smc_particle_sim routine.
Thereafter simulations are calibrated towards the hospital data.
Three possible predictions are made available: Filtering through \eqn{p(x_t|y_{1:t})}, fixed-lag smoothing through \eqn{p(x_t|y_{1:t+u})} and
full smoothing through \eqn{p(x_t|y_{1:T})}. All incidence predictions are only given by full smoothing.
Simulation results related to R and AMP are written into some output files. All predictions are returned as a list.

There are currently many global parameter used in the underlying routine 'run_model'. No control on that
}
\examples{
res.smc = smc_loglik(params=c(R0,Reff,AMP),B=10,dynamic.R=TRUE)
}
