% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_particle_sim.R
\name{smc_particle_sim}
\alias{smc_particle_sim}
\title{Particle simulation of the spread model and log-likelhood evaluation based on hospital data 
Mainly to be used within smc_loglik
Global variables used:
'mobility_matrices', assumed to start at 'input_date'
'hospD'
'testD'}
\usage{
smc_particle_sim(
  input,
  input_date,
  loc.seed,
  mob_mat,
  unit.samp,
  day.numb,
  hospD,
  testD,
  pbar,
  qbar,
  prob_h,
  alpha_hosp,
  alpha_test,
  pi_test,
  MaxdaysH,
  MaxdaysT,
  sim.data = FALSE
)
}
\arguments{
\item{'input'}{A list with 'particle' to start from, 'Rsim', a vector with R-valyes for the differen days 'AMP' and 'inshist'}

\item{'input_date'}{start date for simulation}

\item{'loc-seed'}{seeding for call to run_model, should be adjusted for input_date}

\item{'mob_mat'}{mobility matrix, should be adjusted for input_date}

\item{'unit.samp'}{number of days to simulate}

\item{'day.numb'}{The day number for input_date}

\item{'hospD'}{hospital data, assumed to start at initial date at whole run (not current input_date)}

\item{'testD'}{test data, assumed to start at initial date at whole run (not current input_date)}

\item{'pbar'}{Survival probability for delay until hospitalization}

\item{'qbar'}{Survival probability for delay until testing}

\item{'prob_h'}{Probability for hospitalisation -> NOW A VECTOR WITH LENGTH=NUMBER OF DAYS - NOTE THAT beta_hosp is also a vector then}

\item{'alpha_hosp'}{Parameter in beta-binomial distribution for hospital data}

\item{'alpha_test'}{Parameter in beta-binomial distribution for test data}

\item{'pi_test'}{Parameters describing detection probability}

\item{'MaxdaysH'}{Maximum number of days delay until hospitalization}

\item{'MaxdaysT'}{Maximum number of days delay until testing}

\item{'sim.data'}{If true, observed data are replaced by simulated ones}
}
\value{
A list containing simulated particles of latent variables ('latent'), 'ins' and corresponding log-weights
}
\description{
First particles are 'unit.samp' days from 'input_date'
Thereafter log-likelhood is calculated based on hospital data
}
\details{
There are currently many global parameter used in the underlying routines. No control on that
}
