% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_plot_cmp_median.R
\name{smc_plot_cmp_median_R}
\alias{smc_plot_cmp_median_R}
\title{Plotting routine for estimates of R. Make plots for the given set of quantiles}
\usage{
smc_plot_cmp_median_R(
  mat,
  days,
  col = "#08769c",
  title = "",
  end_date = end_date,
  test_date = NULL,
  pred_date = NULL
)
}
\arguments{
\item{'mat'}{Matrix of R-simulations, rows corresponding to days, columns to particles}

\item{'days'}{Vector of dates, length should correspond to number of rows in 'mat'}

\item{'quant'}{Quantiles to be plotted, default is c(0.95,0.9,0.5,0.01)}

\item{'col'}{Colour code used for the different quantiles. Default is col = "#08769c" giving different blue colors}

\item{'title'}{Optional title to include to the plot. Default is NULL}

\item{'trunc'}{If included, plots are truncated from above along the y-axis for this value}

\item{'mat2'}{An additional alternative set of R-simulations. If included, 50 percent quantiles are added (with green color) on the plot}

\item{'end_date'}{Last date to be plotted.}

\item{'test_date'}{Time for which test data are included. If given, a vertical red line is given at this date}

\item{'pred_date'}{Time for which prediction starts. If given, a vertical red line is given at this date}
}
\value{
An object of class gg / ggplot which can be plotted
}
\description{
Plotting routine for estimates of R. Make plots for the given set of quantiles
}
