% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_PMCMC_par.R
\name{smc_PMCMC_par}
\alias{smc_PMCMC_par}
\title{Particle MCMC routine for updating parameters}
\usage{
smc_PMCMC_par(
  mu_R0 = 3.3,
  sd_R0 = 0.1,
  mu_AMP = 2.8,
  sd_AMP = 0,
  dyn_model = "RW",
  param_dynamicR = list(alpha = 2.4, beta = 0.28, phi.a = 100, phi.b = 900, a0 = 0.5,
    mu0 = 0, kappa0 = 16),
  lambda = 9.08,
  lsize = 3,
  lch = NULL,
  prob_h = c(0.036, 0.022),
  change_prob = c(as.Date("2020-05-01")),
  alpha_hosp = 8,
  alpha_test = 8,
  q_test = c(0, 0.05, 0.08, 0.16, 0.4, 0.3, 0.01),
  pi_test = c(-1.175, 7.4e-05),
  start_date = as.Date("2020-02-17"),
  dynR_date = as.Date("2020-03-08"),
  end_date = as.Date("2020-05-05"),
  pred_date = NULL,
  hosp_data = hospitalisations_df,
  test_data = test_data,
  pop = pop,
  loc.seed = seeding,
  mob_mat = mobility_matrices,
  resdir = "./",
  sig.prop = c(0.2, 1, 1, 1, 0.01),
  samp_ind = 1:3,
  B = 20,
  Bmem = B,
  B2 = NULL,
  unit.samp = 3,
  resamp.frac = 1.1,
  smo.lag = 25,
  parallel = TRUE,
  pmcmc = TRUE,
  VERBOSE = FALSE,
  DEBUG = FALSE,
  output_latent = FALSE,
  seed = 123,
  Nmcmc = 10,
  new.L = Nmcmc + 1,
  N_cores = 4
)
}
\arguments{
\item{start_date}{starting date for initial simulations}

\item{hosp_data}{data frame containing hospidal data, with components date and N.}

\item{B}{Number of particles. Default to 100.}

\item{resamp.frac}{When effective number of samples below resamp.fram, resampling is performed. Default to 1. in which case no resampling is performed.}

\item{parallel}{Logical variable, if true parallel computing is performed.}

\item{pmcmc}{Logical value, if true smc_particle_sim_pmcmc is used for simulation of particles}

\item{DEBUG}{If true, some output are written during running}

\item{seed}{Seed number. Default to 123.}

\item{init_date}{date until particles are simulated before calibration towards data}

\item{sig.R0}{Standard deviation in dynamic random-walk model for. If NULL, static model is used. Default to NULL}
}
\value{
data frame with all iterations including parameters and estimated marginal likelihood
}
\description{
Uses estimates of marginal likelihoods using the smc_loglik routine
Make Gaussian random walk moves on log-scale
Also assumes flat prior on log-scale
So far only using default input to smc_loglik, should be modified
}
