% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_predict_R.R
\name{smc_predict_R_rw}
\alias{smc_predict_R_rw}
\title{Prediction of R based on output from smc_loglik_mem. 
Assumes all inference results are stored in files that are read within the routines.
Parameter values from the last inference time points are used (one set for each particle)
One routine for each model for R.}
\usage{
smc_predict_R_rw(dir, npred)
}
\arguments{
\item{'dir'}{Directory (relative to working directory) for files with simulations of R}

\item{'npred'}{number of days to do prediction}
}
\value{
A matrix containing both the simulated R's before the prediction date (based on fixed-lag smoothing) and the predicted values
}
\description{
Prediction of R based on output from smc_loglik_mem. 
Assumes all inference results are stored in files that are read within the routines.
Parameter values from the last inference time points are used (one set for each particle)
One routine for each model for R.
}
