% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_resamp.R
\name{smc_resamp}
\alias{smc_resamp}
\title{Resampling using}
\usage{
smc_resamp(weights, N = length(weights))
}
\arguments{
\item{weights}{Vector proportional to probabilities}

\item{N}{Number of samples, default to length of weights.}
}
\value{
Vector of samples indices
}
\description{
Resampling using
}
