% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_run_model.R
\name{smc_run_model}
\alias{smc_run_model}
\title{New run_model for smc}
\usage{
smc_run_model(
  mobility_matrices,
  start_date,
  end_date,
  R_0 = 2,
  Reffs = c(),
  AMP_factor = 1.5,
  reff_change_dates = c(),
  latent_period = 2,
  presymptomatic_period = 2,
  infectious_period = 3,
  presymptomatic_relative_infectiousness = 1.3,
  asymptomatic_prob = 0.4,
  asymptomatic_relative_infectiousness = 0.1,
  N_sim = 100,
  seeding = NULL,
  se1e2iiar_pop = NULL,
  parallel = T,
  inputSeed = NULL
)
}
\arguments{
\item{seeding}{-> Data frame including importations. 3 columns: location_code, day (day 1 correspond to start date), n}

\item{se1e2iiar_pop->}{Data frame with the locations. Columns: location_code,S,E1,E2,I,Ia,R.  It defines the number of locations}

\item{N_sim->Number}{of simulations per run (spread package returns the average of all of them)}
}
\description{
New run_model for smc
}
