% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_sim_R.R
\name{smc_sim_R_RW}
\alias{smc_sim_R_RW}
\title{Routine for simulating the dynamic R (different routines for different models)}
\usage{
smc_sim_R_RW(R_sim, Suff_R, unit.samp, param_R)
}
\arguments{
\item{R_sim}{a matrix of dimension (unit.samp+1)*B with the first row giving the current R values.}

\item{Suff_R}{a matrix of dimesion B*nsuff where nsuff is the number of sufficient statistics for updating parameters}

\item{dyn_model}{The model for dynamic R, currently only RW (random walk) is implemented}

\item{param_dynamicR}{List with parameters related to dynamic model for R. For RW, mean and standard deviation in inverse  Gamma distribution for variance in dynamic R distribution. Default is (1,1).}

\item{mu_AMP}{R0 The amplification factor AMP is assumed to follow a lognormal distribution with expectation mu_AMP. Default is 2.8.}

\item{sd_AMP}{The standard deviation in the lognormal distribution for AMP. Default is 0.0 in which case no updating of AMP is performed.}

\item{prob_h}{Probability for hospitalisation. Default to 0.036.}

\item{alpha_hosp}{First parameter in beta-binomial distribution for hospitalization. If NULL, the model reduces to a binomial distribution with 'prob_h' being the probability for hospitalization. In general, the second (beta) parameter is calculated as 'alpha_hosp * (1 - prob_h) / prob_h'. Default is NULL.}

\item{alpha_test}{First parameter in beta-binomial distribution for testdata. If NULL, the model reduces to a binomial distribution with 'prob_h' being the probability for hospitalization. In general, the second (beta) parameter is calculated as 'alpha_hosp * (1 - prob_h) / prob_h'. Default is NULL.}

\item{pi_test}{Detection probability for test data}

\item{lambda}{Expected length from sickness to hospitalization. Default to 9.08.}

\item{lambda_size}{Size parameter in the Negative binomial distribution for length until hospitalization. Default is NULL in which case a Poisson distribution is used.}

\item{lambda_ch}{If 'lambda' and 'lambda_size' are vectors, this correspond to the dates for changes in these parameters}
}
\value{
A list containing particles of latent variables ('latent'), particles of parameters ('param') and estimate of marginal log-likelihood.
}
\description{
Routine for simulating the dynamic R (different routines for different models)
}
